# perf(platform-server): speed up resolution of base